+++ packages/form/src/components/SchemaForm/index.tsx
@@ -2,13 +2,7 @@
 import React, { useCallback, useImperativeHandle, useMemo, useRef, useState } from 'react';
 import type { FormInstance, FormProps } from 'antd';
 
-import {
-  LabelIconTip,
-  omitUndefined,
-  useLatest,
-  runFunction,
-  convertInitialValue,
-} from '@ant-design/pro-utils';
+import { LabelIconTip, omitUndefined, useLatest, runFunction } from '@ant-design/pro-utils';
 import { renderValueType } from './valueType';
 import type { FormSchema, ProFormColumnsType, ProFormRenderValueTypeHelpers } from './typing';
 import omit from 'omit.js';
@@ -104,13 +98,6 @@ function BetaSchemaForm<T, ValueType = 'text'>(props: FormSchema<T, ValueType>) 
               tooltip={originItem.tooltip || originItem.tip}
             />,
           );
-          // 如果指定了 dateFormatter 并且传入的 initialValue 是合法的日期。则也对 initialValue 进行转换
-          let initialValue = originItem.initialValue;
-          initialValue = convertInitialValue(
-            initialValue,
-            props.dateFormatter || false,
-            (originItem.valueType as string) || 'text',
-          );
 
           const item = omitUndefined({
             title,
@@ -121,7 +108,7 @@ function BetaSchemaForm<T, ValueType = 'text'>(props: FormSchema<T, ValueType>) 
             columns: originItem.columns,
             valueEnum: originItem.valueEnum,
             dataIndex: originItem.key || originItem.dataIndex,
-            initialValue,
+            initialValue: originItem.initialValue,
             width: originItem.width,
             index: originItem.index,
             readonly: originItem.readonly,