+++ tests/card/index.test.tsx
@@ -3,7 +3,6 @@ import React from 'react';
 import ProCard from '@ant-design/pro-card';
 import { waitForComponentToPaint } from '../util';
 import { act } from 'react-dom/test-utils';
-import { Grid } from 'antd';
 
 jest.mock('antd/lib/grid/hooks/useBreakpoint');
 
@@ -20,23 +19,6 @@ describe('Card', () => {
       wrapper.find('AntdIcon.ant-pro-card-collapsible-icon').simulate('click');
     });
     expect(fn).toBeCalled();
-  });
-
-  it('🥩 resize breakpoint', async () => {
-    // @ts-ignore
-    Grid.useBreakpoint.mockReturnValue({ xs: true });
-
-    const wrapper = mount(
-      <ProCard
-        style={{ marginTop: 8 }}
-        gutter={[{ xs: 8, sm: 8, md: 16, lg: 24, xl: 32, xxl: 32 }, 16]}
-        title="24栅格"
-      >
-        <ProCard colSpan={{ xs: 2, sm: 4, md: 6, lg: 8, xl: 10, xxl: 12 }}>Col</ProCard>
-      </ProCard>,
-    );
-
-    await waitForComponentToPaint(wrapper);
   });
 
   it('🥩 collapsible defaultCollapsed', async () => {